+++ project/src/components/host-card/index.ts
@@ -1,0 +1,1 @@
+export { default } from './host-card';
