+++ project/src/components/sort-form/index.ts
@@ -1,0 +1,1 @@
+export {default} from './sort-form';
