+++ project/src/components/sort-option/index.ts
@@ -1,0 +1,1 @@
+export { default } from './sort-option';
