--- project/src/components/sort-options/index.ts
@@ -1,1 +1,0 @@
-export { default } from './sort-options';
